requirements.txt
pytest==8.3.2
playwright==1.45.1
pytest-playwright==0.5.1
allure-pytest==2.13.5
assertpy==1.1